periods in names are bad
This is a short sighted feature where names are simplified by dropping
everything after a period from a file name. But real people use periods
in real names sometimes, which breaks things down the line. Short of
painful fixes, a simple hack is to just substitute periods with
underscores and notify user.