better handle empty peak files
Gapped peak files in particular always have track lines, which means we can't use file size as a metric for peak calling success or failure. So we now explicitly count lines as a metric of success. As a bonus, we can print to standard output the number of peaks called.